+++ relay-cardinality/src/limiter.rs
@@ -60,7 +60,7 @@ pub struct Entry<T: CardinalityScope> {
 /// Opaque data structure used by [`CardinalityLimiter`] to track
 /// which buckets have been accepted and rejected.
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug, Hash)]
-pub struct EntryId(pub(crate) usize);
+pub struct EntryId(pub usize);
 
 impl<T: CardinalityScope> Entry<T> {
     /// Creates a new entry.
@@ -70,6 +70,7 @@ impl<T: CardinalityScope> Entry<T> {
 }
 
 /// Represents a from the cardinality limiter rejected [`Entry`].
+#[derive(Debug)]
 pub struct Rejection {
     pub(crate) id: EntryId,
 }