+++ relay-server/src/endpoints/spans.rs
@@ -2,6 +2,7 @@ use axum::extract::{DefaultBodyLimit, Json};
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
 use axum::routing::{post, MethodRouter};
+use bytes::Bytes;
 
 use relay_config::Config;
 use relay_spans::TracesData;
@@ -33,6 +34,11 @@ async fn handle(
     Ok(StatusCode::ACCEPTED)
 }
 
-pub fn route(config: &Config) -> MethodRouter<ServiceState> {
+pub fn route<B>(config: &Config) -> MethodRouter<ServiceState, B>
+where
+    B: axum::body::HttpBody + Send + 'static,
+    B::Data: Send + Into<Bytes>,
+    B::Error: Into<axum::BoxError>,
+{
     post(handle).route_layer(DefaultBodyLimit::max(config.max_span_size()))
 }