+++ relay-server/src/endpoints/store.rs
@@ -143,6 +143,11 @@ async fn handle_get(
     Ok(([(header::CONTENT_TYPE, "image/gif")], PIXEL))
 }
 
-pub fn route(config: &Config) -> MethodRouter<ServiceState> {
+pub fn route<B>(config: &Config) -> MethodRouter<ServiceState, B>
+where
+    B: axum::body::HttpBody + Send + 'static,
+    B::Data: Send,
+    B::Error: Into<axum::BoxError>,
+{
     (post(handle_post).get(handle_get)).route_layer(DefaultBodyLimit::max(config.max_event_size()))
 }