+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**Internal**:
+
+- Use a Lua script and in-memory cache for the cardinality limiting to reduce load on Redis. ([#2849](https://github.com/getsentry/relay/pull/2849))
+
 ## 23.12.0
 
 **Features**: